major refactor: separate all entries from visible
Previously, nimmm rescanned the whole directory when toggling
hidden items or when searching. This is obviously not a good
choice. Now, all entries are stored separately and indications
for which entries are visible are stored. This brings a noticeable
performance improvements for large directories (my nix store for
example) and is just overall better.